[7] Allow ordering of list by rank
We do this using query parameters in the Razor page. A previous approach to hide completed items used JQuery instead; although this may have been preferable here to prevent the need to repeat the query on the server, ultimately this would be working against the way that Razor worked. If time permitted, it would be recommended that the hiding of completed items also be rewritten to use query parameters for consistency.

An assumption has been made here that "lower rank number" = "higher priority". Expectations would have been clarified here, or potentially explanatory text added to the item create/edit page to explain this behaviour if users found it confusing.

The UX on this page is a little clunky; again if time permitted it would be ideal to create a more unified UI that includes consistent action buttons.