[8] Create Gravatar service and use it for user details
An in-memory cache is used for results from the Gravatar service to reduce queries to the API. When a list of users is requested, this list is also deduplicated before the calls to the Gravatar service to further prevent unnecessary calls to this service.

A timeout has also been configured for the Gravatar service; since the MVC page is waiting synchronously for this data, this should prevent the page from taking too long to load. As a further iteration, it would be good to turn this into an async call on the page itself so that the user doesn't have to wait on an external service for this page to load.

The "display name" has been taken as the assumed preferred result based on the requirements; this would usually be clarified before building.